Extract `validateWeb3`
