+++ app/src/marketplace/containers/ProductPage/Hero.jsx
@@ -24,7 +24,7 @@ import Notification from '$shared/utils/Notification'
 import { isAddressWhitelisted } from '$mp/modules/contractProduct/services'
 import useAccountAddress from '$shared/hooks/useAccountAddress'
 import type { ProductId } from '$mp/flowtype/product-types'
-import { validateWeb3 } from '$shared/web3/web3Provider'
+import validateWeb3 from '$utils/web3/validateWeb3'
 import getWeb3 from '$utils/web3/getWeb3'
 import getDefaultWeb3Account from '$utils/web3/getDefaultWeb3Account'
 import routes from '$routes'
