+++ app/src/shared/utils/network.js
@@ -1,4 +1,4 @@
-import { validateWeb3 } from '$shared/web3/web3Provider'
+import validateWeb3 from '$utils/web3/validateWeb3'
 import getWeb3 from '$utils/web3/getWeb3'
 import getConfig from '$shared/web3/config'
 import UnsupportedNetworkError from '$shared/errors/UnsupportedNetworkError'
