+++ app/src/utils/web3/StreamrWeb3.test.js
@@ -1,0 +1,8 @@
+import Web3 from 'web3'
+import StreamrWeb3 from './StreamrWeb3'
+
+describe('StreamrWeb3', () => {
+    it('extends Web3', () => {
+        expect(StreamrWeb3.prototype).toBeInstanceOf(Web3)
+    })
+})
